+++ src/main/kotlin/ru/sadv1r/idea/plugin/ansible/vault/editor/AnsibleConfigReader.kt
@@ -2,9 +2,46 @@ package ru.sadv1r.idea.plugin.ansible.vault.editor
 
 import com.intellij.openapi.util.io.FileUtil
 import com.intellij.util.SystemProperties
+import java.io.BufferedReader
 import java.io.File
 import java.io.FileReader
+import java.io.InputStreamReader
 import java.util.*
+import java.util.concurrent.TimeUnit
+
+fun safeGetPasswordFromFile(vaultId: String?): String? {
+    return try {
+        getPasswordFromFile(vaultId)
+    } catch (e: Exception) {
+        null
+    }
+}
+
+private fun getPasswordFromFile(vaultId: String?): String? {
+    val passwordFilePath = getPasswordFilePath()
+
+    return if (passwordFilePath != null) {
+        val file = File(passwordFilePath)
+
+        return if (file.canExecute()) {
+            try {
+                val process = if (vaultId == null) {
+                    Runtime.getRuntime().exec(file.absolutePath)
+                } else {
+                    Runtime.getRuntime().exec(arrayOf(file.absolutePath, "--vault-id", vaultId))
+                }
+
+                process.waitFor(10, TimeUnit.SECONDS)
+                val reader = BufferedReader(InputStreamReader(process.inputStream, Charsets.UTF_8))
+                reader.readText()
+            } catch (e: Exception) {
+                file.readText(Charsets.UTF_8)
+            }
+        } else {
+            file.readText(Charsets.UTF_8)
+        }
+    } else null
+}
 
 fun getPasswordFilePath(): String? {
     System.getenv("ANSIBLE_VAULT_PASSWORD_FILE")